Added functional and container logic
