+++ src/components/menu-item/container.jsx
@@ -1,7 +1,7 @@
 import { useCallback, useState } from "react"
-import { Counter } from "../counter"
 import { useSelector } from "react-redux"
 import { selectDishById } from "../../redux/features/entities/dishes/selector"
+import { MenuItem } from "./component"
 
 const MENUITEM_CONST = {
     min: 0,
@@ -9,10 +9,10 @@ const MENUITEM_CONST = {
     step: 1
 }
 
-export const MenuItem = ({ id }) => {
+export const MenuItemContainer = ({ id }) => {
     const [amount, setAmount] = useState(0)
 
-    const { name, price, ingredients } = useSelector(state => selectDishById(state, id))
+    const dish = useSelector(state => selectDishById(state, id))
 
     const incrementAmount = useCallback((e) => {
         e.preventDefault()
@@ -24,16 +24,6 @@ export const MenuItem = ({ id }) => {
     }, [amount])
 
     return (
-        <>
-            <p>
-                <strong>{name}</strong> - {price}$
-            </p>
-            <Counter
-                value={amount}
-                increment={incrementAmount}
-                decrement={decrementAmount}
-            />
-            <p>{typeof ingredients === 'object' ? ingredients.join(', ') : ingredients}</p>
-        </>
+        <MenuItem dish={dish} amount={amount} incrementAmount={incrementAmount} decrementAmount={decrementAmount} />
     )
 }
