+++ src/components/user-review/component.jsx
@@ -1,0 +1,12 @@
+export const UserReview = ({ selectedUser, dish }) => {
+    const { rating, text } = dish
+
+    return (
+        <>
+            <p>
+                <strong>{selectedUser.name}</strong>: <u>{rating}</u>
+            </p>
+            <p>{text}</p>
+        </>
+    )
+}
